pyproject.toml
[project]
name = "solana-intel-engine"
version = "0.1.0"
description = "On-Chain Intelligence & Execution Engine for Solana"
requires-python = ">=3.11"
dependencies = [
    "asyncpg==0.29.0",
    "redis==5.0.1",
    "neo4j==5.15.0",
    "aiohttp==3.9.1",
    "python-dotenv==1.0.0",
    "websockets==12.0",
    "backoff==2.2.1",
    "yellowstone-grpc-client==1.5.0",
    "grpcio==1.60.0",
    "fastapi==0.109.0",
    "uvicorn==0.27.0",
    "python-socketio==5.11.0",
    "cryptography==42.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "ingestion", "logic", "execution", "analytics"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
